+++ Expanded Cheat Menu/Code/Menus/HelpFunc.lua
@@ -42,8 +42,8 @@ do -- ModUpload
 		[ChoGGi.id] = true,
 	}
 
-	-- this keeps the check saved per session (true = steam, false = paradox)
-	local upload_to_who = true
+--~ 	-- this keeps the check saved per session (true = steam, false = paradox)
+--~ 	local upload_to_who = true
 	-- true = desktop, false = desktop/console
 	local upload_to_whichplatform = false
 
@@ -73,7 +73,7 @@ do -- ModUpload
 	local hpk_path_working
 
 	-- It's fine...
-	local blank_mod, clipboard, test, steam_upload, para_platform
+	local blank_mod, clipboard, test, steam_upload, para_upload, para_platform
 	local mod, mod_path, upload_image, diff_author, result, choices_len, uploading, msg_popup_id
 	local result_msg, result_title, upload_msg = {}, {}, {}
 	local image_steam = "UI/Common/mod_steam_workshop.tga"
@@ -87,15 +87,15 @@ do -- ModUpload
 		local orig_title
 
 		msg_popup_id = MsgPopup(
-			T(5452, "START"),
-			Strings[302535920000367--[[Mod Upload]]]
+			T(5452--[[START]]),
+			T(302535920000367--[[Mod Upload]])
 		)
 
 		-- always start with fresh table
 		table.clear(mod_params)
 
 		-- add new mod
-		local err, item_id, prepare_worked, prepare_results, existing_mod
+		local err, steam_item_id, para_item_id, prepare_worked, prepare_results, existing_mod
 		if steam_upload then
 			if mod.steam_id ~= 0 then
 				existing_mod = true
@@ -103,10 +103,12 @@ do -- ModUpload
 			-- get needed info for mod
 			prepare_worked, prepare_results = Steam_PrepareForUpload(nil, mod, mod_params)
 			-- mod id for clipboard
-			item_id = mod.steam_id
+			steam_item_id = mod.steam_id
 
-		-- paradox mods
-		else
+		end -- steam upload
+
+		-- para upload
+		if para_upload then
 			-- paradox has some annoyances when it comes to mod titles
 			local paradox_title = diff_mod_titles_paradox[mod.id]
 			if paradox_title then
@@ -138,37 +140,9 @@ do -- ModUpload
 
 			prepare_worked, prepare_results = PDX_PrepareForUpload(nil, mod, mod_params)
 
-			item_id = mod[mod_params.uuid_property]
+			para_item_id = mod[mod_params.uuid_property]
 
-			-- tell paradox users if it needs my library
-			local needs_lib = table.find(mod.dependencies, "id", ChoGGi.id_lib)
-				and Strings[302535920001634--[["This mod requires my library mod (ChoGGi's Library ) < use space on the end when searching for it."]]] .. "\n\n"
-				or ""
-
-			-- add some text to ECM description to hopefully reduce people reporting the mod.
-			if testing then
-
-			if mod.id == ChoGGi.id then
-				mod.description = Strings[302535920000990--[["You need to have a mouse to use this mod."]]] .. "\n"
-					.. needs_lib .. Strings[302535920000887--[["If you get a disabled content restrictions error: It's on Paradox to review/restore it (bug them not me).
-
-If you have any issues with this mod then please send me a bug report instead of reporting the mod.
-https://github.com/ChoGGi/SurvivingMars_CheatMods
-https://discord.gg/bkdjEZU
-https://steamcommunity.com/id/ChoGGi/
-SurvivingMarsMods@choggi.org"]]] .. "\n\n\n" .. mod.description
-			else
-				mod.description = needs_lib .. Strings[302535920000887--[["If you get a disabled content restrictions error: It's on Paradox to review/restore it (bug them not me).
-
-If you have any issues with this mod then please send me a bug report instead of reporting the mod.
-https://github.com/ChoGGi/SurvivingMars_CheatMods
-https://discord.gg/bkdjEZU
-https://steamcommunity.com/id/ChoGGi/
-SurvivingMarsMods@choggi.org"]]] .. "\n\n\n" .. mod.description
-			end
-			end
-
-		end
+		end -- para upload
 
 		-- Issue with mod platform (workshop/paradox mods)
 		if not prepare_worked then
@@ -236,7 +210,7 @@ SurvivingMarsMods@choggi.org"]]] .. "\n\n\n" .. mod.description
     local all_files
 		err, all_files = AsyncListFiles(dest_path, nil, "recursive")
 		if err then
-			err = T{1000753, "Failed creating content package file (<err>)", err = err}
+			err = T{1000753--[[Failed creating content package file (<err>)]], err = err}
 		else
 			-- do this after listfiles so it doesn't include it
 			AsyncCreatePath(pack_path)
@@ -310,23 +284,55 @@ SurvivingMarsMods@choggi.org"]]] .. "\n\n\n" .. mod.description
 			if not test then
 				if steam_upload then
 					result, err = Steam_Upload(nil, mod, mod_params)
-				else
-					local org_mod_description = mod.description
+				end
+
+				local org_mod_description = mod.description
+				if para_upload then
+
+					-- DESC FOR PARA
+					-- tell paradox users if it needs my library
+					local needs_lib = table.find(mod.dependencies, "id", ChoGGi.id_lib)
+						and Strings[302535920001634--[["This mod requires my library mod (ChoGGi's Library) < use space on the end when searching for it."]]] .. "\n\n"
+						or ""
+
+					-- add some text to ECM description to hopefully reduce people reporting the mod.
+					if testing then
+
+						if mod.id == ChoGGi.id then
+							mod.description = Strings[302535920000990--[["You need to have a mouse to use this mod."]]] .. "\n"
+								.. needs_lib .. Strings[302535920000887--[["If you get a disabled content restrictions error: It's on Paradox to review/restore it (bug them not me).
+
+If you have any issues with this mod then please send me a bug report instead of reporting the mod.
+https://github.com/ChoGGi/SurvivingMars_CheatMods
+https://discord.gg/bkdjEZU
+https://steamcommunity.com/id/ChoGGi/
+SurvivingMarsMods@choggi.org"]]] .. "\n\n\n" .. mod.description
+						else
+							mod.description = needs_lib .. Strings[302535920000887--[["If you get a disabled content restrictions error: It's on Paradox to review/restore it (bug them not me).
+
+If you have any issues with this mod then please send me a bug report instead of reporting the mod.
+https://github.com/ChoGGi/SurvivingMars_CheatMods
+https://discord.gg/bkdjEZU
+https://steamcommunity.com/id/ChoGGi/
+SurvivingMarsMods@choggi.org"]]] .. "\n\n\n" .. mod.description
+						end
+					end
+					-- DESC FOR PARA
 
 					-- thanks LukeH (line breaks needed for paradox)
 					-- not bold (what's bold on paradox?)
 					mod.description = mod.description:gsub("\n", "<br>"):gsub("%[b%]", ""):gsub("%[%/b%]", "")
 
 					result, err = PDX_Upload(nil, mod, mod_params)
-					-- shouldn't actually matter, but maybe some weird person will use mod editor along with ECM upload?
-					mod.description = org_mod_description
 				end
+
+				mod.description = org_mod_description
 			end
 		end
 
 		-- uploaded or failed?
 		if err and not blank_mod then
-			local msg = T{1000013, "Mod <ModLabel> was not uploaded! Error: <err>",
+			local msg = T{1000013--[[Mod <ModLabel> was not uploaded! Error: <err>]],
 				ModLabel = mod.title, err = err,
 			}
 			result_msg[#result_msg+1] = msg
@@ -344,32 +350,33 @@ SurvivingMarsMods@choggi.org"]]] .. "\n\n\n" .. mod.description
 				print(Translate("<color ChoGGi_red>" .. msg .. "\n" .. tostring(log_error) .. "</color>"))
 			end
 			if log_error then
-				result_title[#result_title+1] = "\n<color ChoGGi_red>" .. T(1000592, "Error") .. "</color>"
+				result_title[#result_title+1] = "\n<color ChoGGi_red>" .. T(1000592--[[Error]]) .. "</color>"
 				result_msg[#result_msg+1] = log_error
 			end
 		else
 			if batch then
-				print(Translate("<color ChoGGi_green>" .. T(1000015, "Success") .. " " .. mod.title .. "</color>"))
+				print(Translate("<color ChoGGi_green>" .. T(1000015--[[Success]]) .. " " .. mod.title .. "</color>"))
 			end
 			if choices_len == 1 then
-				result_msg[#result_msg+1] = T{1000014, "Mod <ModLabel> was successfully uploaded!",
+				result_msg[#result_msg+1] = T{1000014--[[Mod <ModLabel> was successfully uploaded!]],
 					ModLabel = mod.title,
 				}
-				result_title[#result_title+1] = Translate("<color ChoGGi_green>" .. T(1000015, "Success") .. "</color>")
+				result_title[#result_title+1] = Translate("<color ChoGGi_green>" .. T(1000015--[[Success]]) .. "</color>")
 			else
 				result_title[#result_title+1] = mod.title
 			end
 		end
 
 		-- show id in console/copy to clipboard
-		if not test and item_id then
+		if not test and (steam_item_id or para_item_id) then
 			-- don't copy to clipboard if existing mod or not steam or failed
 			if not existing_mod and steam_upload and clipboard and not err then
-				CopyToClipboard("	\"steam_id\", \"" .. item_id .. "\",")
+				CopyToClipboard("	\"steam_id\", \"" .. steam_item_id .. "\",")
 			end
 
 			local id_str = 1000021--[[Steam ID]]
-			if not steam_upload then
+
+			if para_upload then
 				if para_platform then
 					id_str = 1000772--[[Paradox Desktop UUID]]
 				else
@@ -377,7 +384,15 @@ SurvivingMarsMods@choggi.org"]]] .. "\n\n\n" .. mod.description
 				end
 			end
 
-			print(mod.title, ":<color ChoGGi_orange>", Translate(id_str), "</color>:", item_id)
+			if steam_upload and para_upload then
+				print(mod.title, ":<color ChoGGi_orange>",
+					Translate(1000021--[[Steam ID]]), "</color>:", steam_item_id,
+					"\n<color ChoGGi_orange>", Translate(id_str), "</color>:", para_item_id
+				)
+			else
+				print(mod.title, ":<color ChoGGi_orange>", Translate(id_str), "</color>:", steam_upload and steam_item_id or para_item_id)
+			end
+
 		end
 
 		if not test and not err then
@@ -413,7 +428,8 @@ SurvivingMarsMods@choggi.org"]]] .. "\n\n\n" .. mod.description
 			clipboard = choice.check2
 			test = choice.check3
 			steam_upload = choice.check4
-			para_platform = choice.check5
+			para_upload = choice.check5
+			para_platform = choice.check6
 
 			choices_len = #choices
 
@@ -427,10 +443,14 @@ SurvivingMarsMods@choggi.org"]]] .. "\n\n\n" .. mod.description
 
 					mod = choice.mod
 					mod_path = choice.path
+
 					-- pick logo for upload msg boxes
-					if steam_upload then
+					-- merged for dual upload?
+					if steam_upload and para_upload then
+						upload_image = nil
+					elseif steam_upload then
 						upload_image = image_steam
-					else
+					elseif para_upload then
 						upload_image = image_paradox
 					end
 
@@ -458,14 +478,17 @@ SurvivingMarsMods@choggi.org"]]] .. "\n\n\n" .. mod.description
 						local m_c = 0
 
 						m_c = m_c + 1
-						if steam_upload then
-							upload_msg[m_c] = T{1000012,
-								"Mod <ModLabel> will be uploaded to Steam",
+
+						if steam_upload and para_upload then
+							upload_msg[m_c] = T{302535920001381, "Mod <ModLabel> will be uploaded to Steam/Paradox",
 								ModLabel = mod.title,
 							}
-						else
-							upload_msg[m_c] = T{1000771,
-								"Mod <ModLabel> will be uploaded to Paradox",
+						elseif steam_upload then
+							upload_msg[m_c] = T{1000012--[[Mod <ModLabel> will be uploaded to Steam]],
+								ModLabel = mod.title,
+							}
+						elseif para_upload then
+							upload_msg[m_c] = T{1000771--[[Mod <ModLabel> will be uploaded to Paradox]],
 								ModLabel = mod.title,
 							}
 						end
@@ -473,16 +496,16 @@ SurvivingMarsMods@choggi.org"]]] .. "\n\n\n" .. mod.description
 						m_c = m_c + 1
 						upload_msg[m_c] = "\n\n"
 						m_c = m_c + 1
-						upload_msg[m_c] = Strings[302535920001572--[["<color ChoGGi_red>Pack Warning</color>: Will instantly crash SM when calling it a second time, pack the mod manually to workaround it.
+						upload_msg[m_c] = Translate(302535920001572--[["<color ChoGGi_red>Pack Warning</color>: Will instantly crash SM when calling it a second time, pack the mod manually to workaround it.
 You can also stick the executable in the profile folder to use it instead (<green>no crashing</green>):
-<yellow>%s</yellow>."]]]:format(ConvertToOSPath(hpk_path))
+<yellow>%s</yellow>."]]):format(ConvertToOSPath(hpk_path))
 
 						-- show diff author warning unless it's me
 						if diff_author and not testing then
 							m_c = m_c + 1
 							upload_msg[m_c] = "\n\n"
 							m_c = m_c + 1
-							upload_msg[m_c] = Strings[302535920001263--[["%s is different from your name, do you have permission to upload it?"]]]:format(mod.author)
+							upload_msg[m_c] = Translate(302535920001263--[["%s is different from your name, do you have permission to upload it?"]]):format(mod.author)
 						end
 					end
 
@@ -528,10 +551,10 @@ You can also stick the executable in the profile folder to use it instead (<gree
 						end
 						-- and show msg
 						if ChoGGi.ComFuncs.QuestionBox(
-							Strings[302535920000221--[[Batch Upload mods?]]] .. "\n\n"
+							T(302535920000221--[[Batch Upload mods?]]) .. "\n\n"
 								.. table.concat(titles, ", "),
 							CallBackFunc_BQ,
-							Strings[302535920000221--[[Batch Upload!]]],
+							T(302535920000221--[[Batch Upload!]]),
 							nil,
 							nil,
 							upload_image,
@@ -559,7 +582,7 @@ You can also stick the executable in the profile folder to use it instead (<gree
 			local popups = ChoGGi.Temp.MsgPopups
 			local idx = table.find(popups, "notification_id", msg_popup_id)
 			if idx and ChoGGi.ComFuncs.IsValidXWin(popups[idx]) then
-				popups[idx].idText:SetText(Strings[302535920001453--[[Completed]]])
+				popups[idx].idText:SetText(T(302535920001453--[[Completed]]))
 			end
 
 			local error_msgs = {}
@@ -570,7 +593,7 @@ You can also stick the executable in the profile folder to use it instead (<gree
 			end
 			error_msgs = table.concat(error_msgs)
 
-			local error_text = Strings[302535920000221--[[See log for any batch errors.]]]
+			local error_text = T(302535920000221--[[See log for any batch errors.]])
 			-- only add error msg if single mod
 			if choices_len == 1 then
 				error_text = error_text .. "\n\n" .. error_msgs
@@ -580,7 +603,7 @@ You can also stick the executable in the profile folder to use it instead (<gree
 			print(Translate(error_msgs))
 			ChoGGi.ComFuncs.MsgWait(
 				error_text,
-				Strings[302535920001586--[[All Done!]]],
+				T(302535920001586--[[All Done!]]),
 				upload_image
 			)
 
@@ -596,10 +619,10 @@ You can also stick the executable in the profile folder to use it instead (<gree
 		if not (Platform.steam or Platform.pops) then
 			local msg = Translate(1000760--[[Not Steam]]) .. "/"
 				.. Translate(1000759--[[Not Paradox]])
-			print(Strings[302535920000367--[[Mod Upload]]], ":", msg)
+			print(Translate(302535920000367--[[Mod Upload]]), ":", msg)
 			MsgPopup(
 				msg,
-				Strings[302535920000367--[[Mod Upload]]]
+				T(302535920000367--[[Mod Upload]])
 			)
 			return
 		end
@@ -636,64 +659,75 @@ You can also stick the executable in the profile folder to use it instead (<gree
 			end
 		end
 
-		local _, image_steam_y = MeasureImage(image_steam)
-		local _, image_paradox_y = MeasureImage(image_paradox)
+--~ 		local _, image_steam_y = MeasureImage(image_steam)
+--~ 		local _, image_paradox_y = MeasureImage(image_paradox)
 
 		ChoGGi.ComFuncs.OpenInListChoice{
-			background_image = upload_to_who and image_steam or image_paradox,
+--~ 			background_image = upload_to_who and image_steam or image_paradox,
 			callback = CallBackFunc,
 			items = item_list,
-			title = Strings[302535920000367--[[Mod Upload]]],
-			hint = Strings[302535920001511--[["AsyncPack will CTD the second time you call it, you can use hpk to pack mods ahead of time.
+			title = T(302535920000367--[[Mod Upload]]),
+			hint = T(302535920001511--[["AsyncPack will CTD the second time you call it, you can use hpk to pack mods ahead of time.
 
 https://github.com/nickelc/hpk
 <green>hpk create ""Mod folder"" ModContent.hpk</green>
-Move archive to ""Mod folder/Pack/ModContent.hpk"""]]] .. "\n\n" .. Strings[302535920001572]:format(ConvertToOSPath(hpk_path)),
+Move archive to ""Mod folder/Pack/ModContent.hpk"""]]) .. "\n\n" .. Translate(302535920001572):format(ConvertToOSPath(hpk_path)),
 			height = 800.0,
 			multisel = true,
 			checkboxes = {
-				{title = Strings[302535920001260--[[Blank]]],
-					hint = Strings[302535920001261--[["Uploads a blank mod, and prints id in log."]]],
+				{title = T(302535920001260--[[Blank]]),
+					hint = T(302535920001261--[["Uploads a blank mod, and prints id in log."]]),
 				},
-				{title = Strings[302535920000664--[[Clipboard]]],
-					hint = Strings[302535920000665--[[If uploading a mod this copies the mod's steam id clipboard.]]],
+				{title = T(302535920000664--[[Clipboard]]),
+					hint = T(302535920000665--[[If uploading a mod this copies the mod's steam id clipboard.]]),
 					checked = true,
 				},
 				{title = T(186760604064, "Test"),
 					level = 2,
-					hint = Strings[302535920001485--[[Does everything other than uploading mod to workshop (see AppData/ModUpload).]]],
+					hint = T(302535920001485--[[Does everything other than uploading mod to workshop (see AppData/ModUpload).]]),
 				},
-				{title = upload_to_who and Strings[302535920001506--[[Steam]]] or T(5482, "Paradox"),
+
+
+				{title = T(302535920001506--[[Steam]]),
 					level = 2,
-					hint = Strings[302535920001507--[[Uncheck to upload to Paradox mods (instead of Steam).]]],
-					checked = upload_to_who,
-					func = function(dlg, check)
-						upload_to_who = check
-						if check then
-							dlg.idCheckBox4:SetText(Strings[302535920001506--[[Steam]]])
-							dlg.idCheckBox5:SetVisible()
-							dlg.idBackgroundFrame:SetImage(image_steam)
-							dlg.idBackgroundFrame:SetMinHeight(image_steam_y)
-						else
-							dlg.idCheckBox4:SetText(T(5482, "Paradox"))
-							dlg.idCheckBox5:SetVisible(true)
-							dlg.idBackgroundFrame:SetImage(image_paradox)
-							dlg.idBackgroundFrame:SetMinHeight(image_paradox_y)
-						end
-					end,
-					-- no pops means no sense in showing this
-					visible = Platform.pops,
+					hint = T(302535920001507--[[Upload to Steam Workshop]]),
+					checked = true,
+--~ 					checked = upload_to_who,
+--~ 					func = function(dlg, check)
+--~ 						upload_to_who = check
+--~ 						if check then
+--~ 							dlg.idCheckBox4:SetText(Strings[302535920001506--[[Steam]]])
+--~ 							dlg.idCheckBox5:SetVisible()
+--~ 							dlg.idBackgroundFrame:SetImage(image_steam)
+--~ 							dlg.idBackgroundFrame:SetMinHeight(image_steam_y)
+--~ 						else
+--~ 							dlg.idCheckBox4:SetText(T(5482, "Paradox"))
+--~ 							dlg.idCheckBox5:SetVisible(true)
+--~ 							dlg.idBackgroundFrame:SetImage(image_paradox)
+--~ 							dlg.idBackgroundFrame:SetMinHeight(image_paradox_y)
+--~ 						end
+--~ 					end,
+--~ 					-- no pops means no sense in showing this
+--~ 					visible = Platform.pops,
 				},
-				{title = Strings[302535920001509--[[Platform]]],
+
+				{title = T(5482--[[Paradox]]),
 					level = 2,
-					hint = Strings[302535920001510--[["Paradox mods platform: Leave checked to upload to Desktop only or uncheck to upload to Desktop and Console.
-If you have a uuid in your metadata.lua this checkbox is ignored and it'll try the any uuid then the desktop uuid."]]],
+					hint = T(302535920001662--[[Upload to Paradox Mods]]),
+					checked = true,
+				},
+
+
+				{title = T(302535920001509--[[Platform]]),
+					level = 2,
+					hint = T(302535920001510--[["Paradox mods platform: Leave checked to upload to Desktop only or uncheck to upload to Desktop and Console.
+If you have a uuid in your metadata.lua this checkbox is ignored and it'll try the any uuid then the desktop uuid."]]),
 					checked = upload_to_whichplatform,
 					func = function(_, check)
 						upload_to_whichplatform = check
 					end,
-					-- It defaults to hidden, so if it's paradox then we change it to visible
-					visible = not upload_to_who,
+--~ 					-- It defaults to hidden, so if it's paradox then we change it to visible
+--~ 					visible = not upload_to_who,
 				},
 			},
 		}