+++ Mods ChoGGi/Surface Concrete/metadata.lua
@@ -1,0 +1,25 @@
+return PlaceObj("ModDef", {
+	"title", "Surface Concrete",
+	"id", "ChoGGi_SurfaceConcrete",
+	"steam_id", "2749575462",
+	"pops_any_uuid", "a38801b2-f3a7-4dbe-a660-f5523c465114",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagGameplay", true,
+	"description", [[
+Surface deposits of concrete spawn around actual deposits for some quick concrete without mining.
+
+This will let you see where concrete is before scanning for it.
+
+
+These have been in since release, and apparently never used.
+]],
+})